Organized db functions into separate files for better organization
